MoviePipeline: The UI now allows cvars to be enabled/disabled. Also shows a help tooltip for the cvar to describe what it does, as well as value resolution tooltip for where the cvar value originated (the startup value, the preset value, and the value set in MRQ).
If a cvar is disabled, it will not be included in cvar overrides when rendering, and the value shown in the UI will be the preset value of the cvar (if one exists), else the startup value.

Established a dependency on the ConsoleVariablesEditor. This is initially for gaining access to CommandInfo, but future updates will utilize more of ConsoleVariablesEditor.

Migrated ConsoleVariableSetting::ConsoleVariables away from a TMap to an array of ConsoleVariableEntry structs. This was done to support enable/disable state, as well as support
any additional cvar-related state that needs to be tracked in the future.
* Renamed ConsoleVariables to ConsoleVariables_DEPRECATED.
* ConsoleVariables are converted/merged into the new CVars property upon load.
* If scripting modifies ConsoleVariables after initial load, they are merged into CVars upon cvar apply. Note there is a slight potential for data loss here if a script modifies ConsoleVariables, but then MRQ is not run in PIE (meaning ApplyCVarSettings() is never run). ConsoleVariables will not save due to being deprecated, and it will never be merged into CVars.
* Scripts should use the new cvar mutator methods rather than accessing cvars directly.

#jira UE-170937
#rb matt.hoffman
#preflight 63bc918368068a8bd60cc43c

[CL 23621480 by chad williams in ue5-release-engine-staging branch]