geometry collections should immediately add itself to the external acceleration structure
For large scenes, it may take multiple seconds (if not longer) for a geometry collection to show up in the external acceleration structure causing it to be ignored by movement sweep checks (and others). This is because the geometry collection relies on the internal -> external acceleration structure copy which may take awhile to complete. The single particle physics proxy (i.e. static meshes) will add itself manually to the external acceleration structure in FInitBodiesHelperBase::InitBodies via FChaosScene::AddActorsToScene_AssumesLocked.

#preflight 63bdbdadc45a2c81e01b5813
#rb vincent.robert

[CL 23632331 by michael bao in ue5-release-engine-staging branch]