Niagara Data Channels - Supporting CL
This CL contains supporting code for an upcoming feature, Niagara Data Channels.
Further supporting CLs will follow and finally the main CL which contains the feature.

- Niagara DI functions can now have optionally variable numbers of input & output parameters.
DIs can mark their function signatures to allow this or not.
DIs must handle these variable parameters in their function binding and hlsl gen etc.

- DIs can now provide context menu actions on their function call nodes.
This will allows DataChannel function calls to init themselves with the right Inputs and Outputs but could be broadly useful for a number of DI functions.

- New HLSL generation API for DIs.
This passes in a context object from the translator so that we can more easily update things in future without 1000 boiler plate changes.
To avoid said changes this time, I'm having the base implementation of the new API call into the existing API with the relevant parts of the context.
No DI or it's parent chain should implement both APIs.

 - VM DI Info and GPU function info in NiagaraShadrers now track these variadic inputs.
This has a few knock on changes. One being a custom verison bump, the other being a common refernce struct for FNiagaraVariableBase in NiagaraCore.
Required so that code in NiagaraShader could serialize a refernce to an FNiagaraVariableBase. Actually moving this class to NiagaraCore is a headache of knock on dependencies.

- NiagaraShaders now serialize their base compile hash. This allows DIs to generate per-script intermediate parameter data and then look this up for setting shader parameters correctly.

- Bumping DDC and Compiler guids.


#jira UE-153996
#rb Stu.Mckenna, Rob.Krajcarski
#preflight none

[CL 23650985 by rob krajcarski in ue5-release-engine-staging branch]