nDisplay: Added shared cross GPU media io.
This allows for sharing viewports between render nodes with zero or 1 frame latency. This is an alternative to mGPU that in many cases can be more efficient. It should be used w/o SLI enabled. Two nDisplay nodes would run on the same PC, one with r.GraphicsAdapter 0 and the other with r.GraphicsAdapter 1, which can be set in the nDisplay config, and overridden from a newly added option in Switchboard.

#jira
#rb simon.therriault, andrey.yamashev
#preflight 63c0b3d6b06522475052b25d

[CL 23673035 by alejandro arango in ue5-release-engine-staging branch]