Iris - MinimalGameplayCueReplicationProxy support.
Implemented NetSerializer and ReplicationFragment to mimic MinimalGameplayCueReplicationProxy behavior.

Note: It's advised not to implement custom gameplay logic in a NetSerialize method. It's adviced to markup struct members with UPROPERTY when they're to be replicated, even if there's a custom NetSerialize method. It's adviced to markup struct members with UPROPERTY(NotReplicated) when a UPROPERTY shouldn't be replicated, even when there's a custom NetSerialize method.

#rb mattias.hornlund
#jira UE-169147
#preflight 63c1213702024f93d8f8e171

[CL 23675719 by peter engstrom in ue5-release-engine-staging branch]