Always set an output path for UBT regardless of the configuration/platform it is built for
The "dotnet build" command run by BuildUBT.bat does not specify a value for the platform property. MSBuild can automatically select "x64" as a platform, in particular if the VS Dev Command prompt is used or the user has run any of the vcvars.bat as these will set a "Platform" environment variable. As there are no property groups for this platform in UnrealBuildTool.cs, the build result was not ending up where RunUBT.bat expected it to be. This was manifesting as either a stale Binaries/DotNet/UnrealBuildTool/UnrealBuildTool.dll, or no DLL at all.

#rb joek
#rnx
#preflight 63c79b34fb1a8cf24591c7d5

[CL 23752621 by martin ridgers in ue5-release-engine-staging branch]