Removed old stale code that was preventing LightingOnly from working correctly with static meshes.
The removed condition was forbidding meshes with materials that add some form of displacement to support LightingOnly mode. The comment indicates that the intention was to forbid meshes with tessellation enabled. Since tessellation is completely been removed, the only two other cases that may mark the material as modifying mesh position is either WorldPositionOffset or PixelDepthOffset. Without the if, meshes with materials that write to either these two attributes will now support LightingOnly, although they will lose any displacements when this view mode is on.

#rb josie.yang
#preflight 63c8112b0225f00e1413c85f

[CL 23770988 by massimo tristano in ue5-release-engine-staging branch]