Fix calculation of temporary texture upload resource in D3D12 RHI
RowSizesInBytes was used to compute sizes of the first mip and the tail and to allocate upload resource memory. This can result in under-estimation, as the row size in bytes (i.e. valid data range) may be smaller than the row pitch (i.e. number of bytes between rows).

The pitch is what's used inside MemcpySubresource (called from D3DX12 UpdateSubresources) to compute the destination address. Since the smaller memory footprint was previously computed, smaller buffer range is sub-allocated from FD3D12FastAllocator and adjacent sub-allocations could be corrupted by MemcpySubresource.

The solution is to just use the total size of the resource (including all mips) returned by GetCopyableFootprints() and the offset of mip1 to compute the size of mip0 and the total size of the tail.

#preflight 63ca1bb5977c626356e17475
#rb mihnea.balta

[CL 23791638 by yuriy odonnell in ue5-release-engine-staging branch]