[PCG] First refinement of crc caching system - evaluate crc on point data, and default to not using dependencies crc as output crc.
Previously dependencies crc was used as both a caching lookup key and an output crc to pass to downstream nodes. The issue with this is that regeneration currently produces new crcs for input data (regardless of whether the data itself actually changed), and long term it may always do this (we may never crc all landscape data, so it may always be approximated with a counter-based UID). Propagating such spuriously changing crcs downstream leads to unnecessary executions.

To combat this, this change enables accurate crc of point data, so 'spurious' crcs will be replaced with precise crcs once data has collapsed to points. This reduced regenerating a heavy scene (warm cache) from 43s to 18s.

Crc'ing itself is not cheap, so the Crc value is cached within the PCGData to avoid recomputing when data passed from node to node. This reduced the regenerate time further from 18s to 12s. Overall 3.6x speedup for warm cache (as would be the case while authoring).

Final note - I expected the metadata crc to be costly, but in my tests its a tiny proportion of the crc'ing time (4% of total crc time in metadata). This needs to be monitored going forwards.

#preflight 63d80cf5ec3d1af440a2c3f1
#rb julien.lheureux

[CL 23925019 by huw bowles in 5.2 branch]