Unshelved from pending changelist '23946517':
Fixed GC component implicit generation calculating incorrect sizes for box clusters.

We generate box cluster implicits by wrapping a bounding box around all child verts during DDC data generation, however we were appending each child's verts to a list, then transforming the verts and adding them all to the bounds. This effectively means that for child 0 of a cluster of n nodes we would apply the cluster-to-child transform of every child 1...n to the verts of child 0 - leading to wildly incorrect bounds after calculation. This was also leading to the cluster immediately breaking on the physics thread if simulating imparting non-physical impulses to the collection

#rb Cedric.Caillaud
#jira UE-174010
#preflight 63d957b0b73528c8f71f60be

[CL 23948622 by jessica agee in 5.2 branch]