Runtime/Mac: Undefine PF_MAX so it can be defined by UE for pixelformat instead of socket.h
With XCode14.x, there were changes made to PrintCore.h to include networking code. When we include Cocoa.h, we end up pulling this code in via a long include chain. This results in now pulling in socket.h when we don't need it and PF_MAX getting defined by socket.h instead of by UE.

#preflight https://horde.devtools.epicgames.com/job/63dc54ff244dc45a20776e42
#jira none

[CL 24041796 by zack neyland in 5.2 branch]