Only run FSpanAllocator::Consolidate when no free space is immediately available
It is possible for there to be freed spans that are pending consolidation, however no consolidation can actually occur. In this case every Allocate() will still pay the cost of looking for blocks to consolidate, making it much more expensive.
Simple solution is to only run consolidation if there is no available space otherwise. The downside is that this can theoretically result in worse fragmentation over time.

Significantly improves FLumenSceneData::ProcessLumenSurfaceCacheRequests performance in some situations.

#preflight 63fd0a28ef1b24bf94b30163
#jira none
#rb ola.olsson

[CL 24430890 by Yuriy ODonnell in 5.2 branch]