USD: Also fetch "used material" info by reading the relationship directly, even if no material binding schema is present.
#jira UE-178065
#preflight 63ff7f37c63aaf1e13fd0074

In the USD version we have in 5.3 when a prim just has the material:binding relationship but not the MaterialBindingAPI schema USD itself emits a warning saying "hey these material bindings are invalid since the prim doesn't have the schema".

In the USD version we have in 5.2 we don't get such warning... This is an issue because the new mechanism we added to only parse materials used by the current material purpose would consider a "valid assignment" only one done with the schema. This means that if you're missing the schema we'd ignore the binding (as if it was from an unused material purpose), not generated the materials and there would be no feedback at all.

I'm not sure if no warning means it's an allowed thing to do or not, but we have always parsed material assignments like that anyway (only this filtering of "used materials" that was exclusively checking via the schema). I think we should probably support this in 5.2, even if only due to the fact that if the user is missing these schemas on their prims, he will get no materials parsed and zero feedback about it.

[CL 24463989 by daniel coelho in 5.2 branch]