Switchboard: Fix issue where sometimes ndisplay devices would not be populated in the dropdown.
The issue happens when there are .ndisplay files in the content folders. These are found but not assigned a class and the list of classes is too short, causing an exception later on and aborting populating the dropdown. Now .ndisplay items also get an ndisplay class assigned in the list.

#jira UE-179771
#rb trivial
#preflight python

[CL 24588921 by Alejandro Arango in 5.2 branch]